# eth/tracers: improve tracing performance