@emadami3060 2021-06-08
25800

در تاریخ سه‌شنبه ۸ ژوئن ۲۰۲۱،‏ ۱۶:۲۴ Martin Holst Swende <
***@***.***> نوشت:

> Further improvements:
>
> name                old time/op    new time/op    delta
>
> TransactionTrace-6    28.1ms ±13%    22.7ms ± 8%  -19.07%  (p=0.000 n=9+9)
>
>
>
> name                old alloc/op   new alloc/op   delta
>
> TransactionTrace-6    8.98MB ±12%    7.14MB ±14%  -20.55%  (p=0.000 n=10+10)
>
>
>
> name                old allocs/op  new allocs/op  delta
>
> TransactionTrace-6      163k ± 0%       82k ± 0%  -49.98%  (p=0.002 n=8+10)
>
>
> Total:
>
> name                old time/op    new time/op    delta
>
> TransactionTrace-6     100ms ±12%      23ms ± 8%  -77.36%  (p=0.000 n=10+9)
>
>
>
> name                old alloc/op   new alloc/op   delta
>
> TransactionTrace-6    40.1MB ± 4%     7.1MB ±14%  -82.18%  (p=0.000 n=10+10)
>
>
>
> name                old allocs/op  new allocs/op  delta
>
> TransactionTrace-6      571k ± 0%       82k ± 0%  -85.71%  (p=0.000 n=10+10)
>
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/ethereum/go-ethereum/pull/23016#issuecomment-856698816>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ATNR2PBQIGIX5FKG7IUDOLDTRYAIJANCNFSM46JYHU3Q>
> .
>