@holiman 2021-06-11
I ran this against [this](https://etherscan.io/tx/0x87ba373aa69ee675ae764b49c3d179d2367dd0d970c36b9177dc67f00603b4e9) transaction, 

```
#!/bin/bash

time ./build/bin/geth attach --exec "debug.traceTransaction('0x87ba373aa69ee675ae764b49c3d179d2367dd0d970c36b9177dc67f00603b4e9', {'disableMemory':true, 'disableStorage':true})" 
```
It took about `52s`, and `27s` with stack disabled. 
I also took a cpu profile when running it with stack enabled: 
![profile001](https://user-images.githubusercontent.com/142290/121702317-0e7dcb00-cad2-11eb-85e1-3ad73f0df577.png)