@MariusVanDerWijden 2021-06-30
So I tried this on a really heavy tx we have on ropsten
` time ./build/bin/geth --ropsten attach --exec "debug.traceTransaction('0xcafaf3669df640365da47527902e3e4dff48bd334ab0c5c894dae5ba1f82df27', {'disableMemory':true, 'disableStorage':true, reexec:10000})"`

This PR: 
(60 seconds of it is state regeneration)
```
Executed in  281.69 secs    fish           external
   usr time  151.47 secs  323.00 micros  151.47 secs
   sys time  103.44 secs   67.00 micros  103.44 secs
```

Current Master: 
(59 seconds of it is state regeneration)
(I killed the node, call didn't finish)
```
Executed in  269.17 mins    fish           external
   usr time    6.21 secs  358.00 micros    6.21 secs
   sys time    0.50 secs   77.00 micros    0.50 secs
```