@holiman 2021-06-30
> So I tried this on a really heavy tx we have on ropsten

Looks like it worked then? Can the PR haz an approve?