@holiman 2021-07-01
> 0xcafaf3669df640365da47527902e3e4dff48bd334ab0c5c894dae5ba1f82df27

18M gas on that transactions -- yeah might be a bit slow :)